Fixed front panel brightness indicator scale + random color button.
The indicator was not halfway when using 0.5 as the output value.

Also added an example of how the output could be wrapped in a
light component, to expose it to Home Assistant as a monochromatic
light. Not that useful probably, but a nice display of ESPHome's
flexibility.

The example.yaml now also updates the brightness indicator when
the light brightness changes. When switching the device to night
light mode, it will turn off the front panel illumation completely.

And last but not least: the example configuration now shows a way
to set the device to a random color when touching the color button.