Update technical docs for updated ESP-IDF support
ESP-IDF has an option these days to suppress MAC address CRC errors, and
ESPHome has implemented an advanced configuration flag for ESP32 based
on this.

This patch removes the old information about building a custom Arduino
library for making the lamp work, and replaces it with up-to-date
information about using ESP-IDF instead. The big advantage of using
ESP-IDF here, instead of Arduino, is that it compiles out-of-the-box
with ESPHome and no trickery is required.